refactor(image section): added image section
- extracted image logic from pois eventrecords and tours
